Horrifying JS2 Tab Width Hack
This is what happens when you try to code on an airplane without
internet access. I should circle back on this and find the "canonical"
way of doing this because this feels wrong.

That said:

js2-mode tries to be clever. You cannot simply set js2-basic-offset to 2
and have tab widths of 2. Instead, `setq` calls are treated as
buffer-local, preventing a global change by methods I could figure out
off-hand.

My solution is to add a hook that runs setq for every new js2-mode
buffer. In this manner, we force a tab width of 2.